requirements.txt
Django==4.0.5
djangorestframework
djangorestframework-simplejwt
django-cors-headers==3.12.0
django-environ==0.8.1
psycopg2==2.9.3
Pillow==9.1.1
